+++ settings.gradle.kts
@@ -12,7 +12,11 @@ include(
     "system:article:application",
     "system:article:adapter:persistence:article-rds",
     "system:article:adapter:http-api",
-    "system:user",
+    "system:article:adapter:acl-user",
+    "system:user:domain",
+    "system:user:application",
+    "system:user:adapter:persistence:user-rds",
+    "system:user:adapter:http-api",
     "system:category",
     "system:imgupload",
     "system:comment",
