+++ system/article/adapter/acl-user/build.gradle.kts
@@ -1,0 +1,5 @@
+
+dependencies {
+    implementation(project(":system:user:application"))
+    implementation(project(":system:article:application"))
+}
