+++ system/article/adapter/acl-user/src/main/kotlin/kr/co/jiniaslog/article/adapter/acl/user/UserAcl.kt
@@ -1,0 +1,13 @@
+package kr.co.jiniaslog.article.adapter.acl.user
+
+import kr.co.jiniaslog.article.application.port.UserServiceClient
+import kr.co.jiniaslog.article.domain.UserId
+import kr.co.jiniaslog.lib.context.AntiCorruptLayer
+import kr.co.jiniaslog.user.application.UserServiceStub
+
+@AntiCorruptLayer
+class UserAcl(
+    private val userServiceStub: UserServiceStub,
+) : UserServiceClient {
+    override fun findUserById(id: Long) = UserId(userServiceStub.stub(id).id)
+}
