--- system/article/adapter/persistence/article-rds/src/main/kotlin/kr/co/jiniaslog/article/adapter/http/adapter/persistence/article/JpaArticleRepository.kt
@@ -1,5 +1,0 @@
-package kr.co.jiniaslog.article.adapter.http.adapter.persistence.article
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface JpaArticleRepository : JpaRepository<ArticlePM, Long>
