+++ system/article/application/src/main/kotlin/kr/co/jiniaslog/article/application/port/ArticleIdGenerator.kt
@@ -1,6 +1,6 @@
 package kr.co.jiniaslog.article.application.port
 
-import kr.co.jiniaslog.article.adapter.http.domain.ArticleId
+import kr.co.jiniaslog.article.domain.ArticleId
 
 interface ArticleIdGenerator {
     fun generate(): ArticleId
