+++ system/article/application/src/main/kotlin/kr/co/jiniaslog/article/application/port/UserServiceClient.kt
@@ -1,0 +1,7 @@
+package kr.co.jiniaslog.article.application.port
+
+import kr.co.jiniaslog.article.domain.UserId
+
+interface UserServiceClient {
+    fun findUserById(id: Long): UserId?
+}
