+++ system/article/application/src/main/kotlin/kr/co/jiniaslog/article/application/usecase/ArticleEditUseCase.kt
@@ -1,8 +1,8 @@
 package kr.co.jiniaslog.article.application.usecase
 
-import kr.co.jiniaslog.article.adapter.http.domain.ArticleId
-import kr.co.jiniaslog.article.adapter.http.domain.CategoryId
-import kr.co.jiniaslog.article.adapter.http.domain.TagId
+import kr.co.jiniaslog.article.domain.ArticleId
+import kr.co.jiniaslog.article.domain.CategoryId
+import kr.co.jiniaslog.article.domain.TagId
 
 interface ArticleEditUseCase {
     fun editArticle(articleEditCommand: ArticleEditCommand)
