+++ system/article/application/src/main/kotlin/kr/co/jiniaslog/article/application/usecase/ArticlePostUseCase.kt
@@ -1,15 +1,15 @@
 package kr.co.jiniaslog.article.application.usecase
 
-import kr.co.jiniaslog.article.adapter.http.domain.CategoryId
-import kr.co.jiniaslog.article.adapter.http.domain.TagId
-import kr.co.jiniaslog.article.adapter.http.domain.WriterId
+import kr.co.jiniaslog.article.domain.CategoryId
+import kr.co.jiniaslog.article.domain.TagId
+import kr.co.jiniaslog.article.domain.UserId
 
 interface ArticlePostUseCase {
     fun postArticle(articleCreateCommand: ArticlePostCommand): Long
 }
 
 data class ArticlePostCommand(
-    val writerId: WriterId,
+    val userId: UserId,
     val title: String,
     val content: String,
     val thumbnailUrl: String,
