+++ system/article/application/src/test/kotlin/kr/co/jiniaslog/article/adapter/http/application/ArticlePostUseCaseTests.kt
@@ -4,17 +4,17 @@ import io.kotest.core.spec.style.BehaviorSpec
 import io.kotest.matchers.shouldBe
 import io.mockk.every
 import io.mockk.mockk
-import kr.co.jiniaslog.article.adapter.http.domain.Article
-import kr.co.jiniaslog.article.adapter.http.domain.ArticleFactory
-import kr.co.jiniaslog.article.adapter.http.domain.ArticleId
-import kr.co.jiniaslog.article.adapter.http.domain.CategoryId
-import kr.co.jiniaslog.article.adapter.http.domain.TagId
-import kr.co.jiniaslog.article.adapter.http.domain.WriterId
 import kr.co.jiniaslog.article.application.ArticleService
 import kr.co.jiniaslog.article.application.infra.TransactionHandler
 import kr.co.jiniaslog.article.application.port.ArticleIdGenerator
 import kr.co.jiniaslog.article.application.port.ArticleRepository
 import kr.co.jiniaslog.article.application.usecase.ArticlePostCommand
+import kr.co.jiniaslog.article.domain.Article
+import kr.co.jiniaslog.article.domain.ArticleFactory
+import kr.co.jiniaslog.article.domain.ArticleId
+import kr.co.jiniaslog.article.domain.CategoryId
+import kr.co.jiniaslog.article.domain.TagId
+import kr.co.jiniaslog.article.domain.UserId
 
 internal class ArticlePostUseCaseTests : BehaviorSpec() {
     private val articleRepository: ArticleRepository = mockk(relaxed = true)
@@ -28,7 +28,7 @@ internal class ArticlePostUseCaseTests : BehaviorSpec() {
         Given("다음과 같은 command가 주어질 때") {
             val tags = setOf(TagId(1), TagId(2), TagId(3))
             val command = ArticlePostCommand(
-                writerId = WriterId(1),
+                userId = UserId(1),
                 title = "title",
                 content = "content",
                 thumbnailUrl = "thumbnailUrl",
@@ -43,7 +43,7 @@ internal class ArticlePostUseCaseTests : BehaviorSpec() {
             }
             every {
                 articleFactory.newOne(
-                    writerId = command.writerId,
+                    userId = command.userId,
                     title = command.title,
                     content = command.content,
                     thumbnailUrl = command.thumbnailUrl,
