--- system/article/domain/src/main/kotlin/kr/co/jiniaslog/article/adapter/http/domain/CategoryId.kt
@@ -1,6 +1,0 @@
-package kr.co.jiniaslog.article.adapter.http.domain
-
-import kr.co.jiniaslog.lib.context.ValueObject
-
-@ValueObject
-data class CategoryId(val value: Long)
