+++ system/article/domain/src/main/kotlin/kr/co/jiniaslog/article/domain/CategoryId.kt
@@ -1,0 +1,6 @@
+package kr.co.jiniaslog.article.domain
+
+import kr.co.jiniaslog.lib.context.ValueObject
+
+@ValueObject
+data class CategoryId(val value: Long)
