+++ system/local-lib/src/main/kotlin/kr/co/jiniaslog/lib/context/AntiCorruptLayer.kt
@@ -1,0 +1,5 @@
+package kr.co.jiniaslog.lib.context
+
+@Target(AnnotationTarget.CLASS)
+@Retention(AnnotationRetention.RUNTIME)
+annotation class AntiCorruptLayer
