+++ system/user/application/build.gradle.kts
@@ -1,0 +1,7 @@
+val jar: Jar by tasks
+jar.enabled = true
+jar.archiveFileName.set("${project.name}.jar")
+
+dependencies {
+    api(project(":system:user:domain"))
+}
