+++ system/user/application/src/main/kotlin/kr/co/jiniaslog/user/application/UserServiceStub.kt
@@ -1,0 +1,18 @@
+package kr.co.jiniaslog.user.application
+
+import kr.co.jiniaslog.lib.context.UseCaseInteractor
+import kr.co.jiniaslog.user.domain.User
+
+@UseCaseInteractor
+class UserServiceStub {
+
+    fun stub(id: Long): User {
+        return User(
+            id = id,
+            name = "name",
+            email = "email",
+            password = "password",
+            roles = setOf("USER"),
+        )
+    }
+}
