+++ system/user/domain/src/main/kotlin/kr/co/jiniaslog/user/domain/User.kt
@@ -1,0 +1,10 @@
+package kr.co.jiniaslog.user.domain
+
+// stub
+class User(
+    val id: Long,
+    val name: String,
+    val email: String,
+    val password: String,
+    val roles: Set<String>,
+)
